requirements.txt
flask
flask-cors
pyyaml
pyudev
yara-python
